requirements.txt
click==6.2
ordereddict==1.1
python-dotenv==0.3.0
requests==2.9.1
RPi.GPIO==0.5.11
wheel==0.24.0
